refactor: 패키지 분리 및 패키지명 수정
