+++ src/main/java/controller/BlackjackController.java
@@ -1,17 +1,17 @@
 package controller;
 
-import domain.BlackjackDeck;
-import domain.BlackjackDeckGenerator;
-import domain.BlackjackGame;
-import domain.BlackjackResult;
-import domain.BlackjackWinner;
-import domain.Dealer;
-import domain.DealerWinStatus;
-import domain.Player;
-import domain.TrumpCard;
-import domain.WinStatus;
+import domain.blackjackgame.BlackjackDeck;
+import domain.blackjackgame.BlackjackDeckGenerator;
+import domain.blackjackgame.BlackjackGame;
+import domain.blackjackgame.BlackjackResult;
+import domain.blackjackgame.BlackjackWinner;
+import domain.blackjackgame.DealerWinStatus;
+import domain.blackjackgame.TrumpCard;
+import domain.blackjackgame.WinStatus;
+import domain.participant.Dealer;
+import domain.participant.Player;
 import domain.strategy.BlackjackDrawStrategy;
-import except.BlackJackException;
+import exception.BlackJackException;
 import java.util.List;
 import java.util.Map;
 import java.util.function.Supplier;
