+++ src/main/java/domain/blackjackgame/BlackjackDeckGenerator.java
@@ -1,10 +1,10 @@
-package domain;
+package domain.blackjackgame;
 
 import domain.strategy.DrawStrategy;
+import java.util.ArrayDeque;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashSet;
-import java.util.LinkedList;
 import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
@@ -17,8 +17,10 @@ public class BlackjackDeckGenerator {
         for (Suit suit : suits) {
             trumpCards.addAll(trumpCards(suit));
         }
-        LinkedList<TrumpCard> generatedDeck = new LinkedList<>(trumpCards);
+        LinkedList
+        ArrayDeque<TrumpCard> generatedDeck = new ArrayDeque<>(trumpCards);
         Collections.shuffle(generatedDeck);
+
         return new BlackjackDeck(generatedDeck, drawStrategy);
     }
