+++ src/main/java/domain/blackjackgame/BlackjackResult.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.blackjackgame;
 
 import java.util.List;
