+++ src/main/java/domain/blackjackgame/CardValue.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.blackjackgame;
 
 import java.util.Arrays;
 import java.util.List;
