+++ src/main/java/domain/blackjackgame/DealerWinStatus.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.blackjackgame;
 
 public record DealerWinStatus(int win, int lose) {
 }
