+++ src/main/java/domain/blackjackgame/Suit.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.blackjackgame;
 
 public enum Suit {
