+++ src/main/java/domain/blackjackgame/TrumpCard.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.blackjackgame;
 
 public record TrumpCard(Suit suit, CardValue cardValue) {
