+++ src/main/java/domain/blackjackgame/WinStatus.java
@@ -1,7 +1,7 @@
-package domain;
+package domain.blackjackgame;
 
 public enum WinStatus {
-    
+
     WIN("승"),
     LOSE("패"),
     DRAW("무승부");
