+++ src/main/java/domain/participant/Dealer.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.participant;
 
 public class Dealer extends BlackjackParticipant {
