--- src/main/java/except/BlackJackException.java
@@ -1,7 +1,0 @@
-package except;
-
-public class BlackJackException extends IllegalStateException{
-    public BlackJackException(String message){
-        super(message);
-    }
-}
