+++ src/main/java/exception/BlackJackException.java
@@ -1,0 +1,7 @@
+package exception;
+
+public class BlackJackException extends IllegalStateException {
+    public BlackJackException(String message) {
+        super(message);
+    }
+}
