+++ src/test/java/BlackjackGameTest.java
@@ -2,16 +2,16 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIterable;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import domain.BlackjackDeck;
-import domain.BlackjackDeckGenerator;
-import domain.BlackjackGame;
-import domain.BlackjackResult;
-import domain.CardValue;
-import domain.Dealer;
-import domain.Suit;
-import domain.TrumpCard;
+import domain.blackjackgame.BlackjackDeck;
+import domain.blackjackgame.BlackjackDeckGenerator;
+import domain.blackjackgame.BlackjackGame;
+import domain.blackjackgame.BlackjackResult;
+import domain.blackjackgame.CardValue;
+import domain.blackjackgame.Suit;
+import domain.blackjackgame.TrumpCard;
+import domain.participant.Dealer;
 import domain.strategy.BlackjackDrawStrategy;
-import except.BlackJackException;
+import exception.BlackJackException;
 import java.util.Deque;
 import java.util.LinkedList;
 import java.util.List;
